[daemon/daemonapi] Fix write server sent event (multi line / half write)
This commit fixes writeEvents:
- multi data lines server sent events:
   data: line1
   line2
  must be:
   data: line1
   data: line2

- Add missing write calls for not full writes

- messages now define 'Event() string' instead of 'EventSource() string' to retrieve event field value
  use this to drop publications that are not events

+++ daemon/daemonapi/post_object_subscription.go
@@ -18,14 +18,14 @@ import (
 func (a *DaemonApi) PostObjectSubscription(w http.ResponseWriter, r *http.Request, params PostObjectSubscriptionParams) {
 	var (
 		handlerName = "PostObjectSubscription"
-		limit       int64
+		limit       uint64
 		maxDuration = 5 * time.Second
 	)
 	log := getLogger(r, handlerName)
 	log.Debug().Msg("starting")
 	defer log.Debug().Msg("done")
 	if params.Limit != nil {
-		limit = *params.Limit
+		limit = uint64(*params.Limit)
 	}
 	if params.Duration != nil {
 		if v, err := converters.Duration.Convert(*params.Duration); err != nil {
